requirements.txt
pymupdf>=1.23.0
pypdf>=3.0.0
PyQt6>=6.6.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
pyinstaller>=6.0.0
